Split Code related to storage, array, disk and device to respected blueprints
